% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-package.r
\docType{package}
\name{string}
\alias{string}
\alias{string-package}
\title{string: Make String Processing Easier}
\description{
This package is complementary to stringr package. It includes thress
functions: string_insert, string_cut and string_replace. string_insert
inserts a string to another string at a given position. string_cut cuts a
substring between two given positions from a string. string_replace replaces
a substring between two given positions in a string with another string.
}

