% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_cut.R
\name{string_cut}
\alias{string_cut}
\title{Cut a substring between two given positions from a string}
\usage{
string_cut(a, startPos, endPos)
}
\arguments{
\item{a}{character}

\item{startPos}{integer}

\item{endPos}{integer}
}
\value{
character
}
\description{
Cut a substring between two given positions from a string
}
\examples{
string_cut("apple banana pear", as.integer(7), as.integer(13))
}

