% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_insert.R
\name{string_insert}
\alias{string_insert}
\title{Insert a string to another string}
\usage{
string_insert(a, b, pos)
}
\arguments{
\item{a}{character}

\item{b}{character}

\item{pos}{integer}
}
\value{
character
}
\description{
Insert a string to another string at a given position
}
\examples{
string_insert("apple pear", "banana ", as.integer(6))
}

