% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_replace.R
\name{string_replace}
\alias{string_replace}
\title{Replace a substring between two given positions in a string with another string}
\usage{
string_replace(a, b, startPos, endPos)
}
\arguments{
\item{a}{character}

\item{b}{character}

\item{startPos}{integer}

\item{endPos}{integer}
}
\value{
character
}
\description{
Replace a substring between two given positions in a string with another string
}
\examples{
string_replace("apple banana pear", "peach", as.integer(7), as.integer(12))
}

